LayoutTests/js/dfg-to-string-side-effect-clobbers-toString-expected.txt

Tests what happens when you do ToString twice, and it has a side effect that clobbers the toString method in between the two ToStrings.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
PASS foo(new String("hello"), sideEffect) is "hellohello"
hi!
PASS foo(new String("hello"), sideEffect) is "hello150"
hi!
PASS foo(new String("hello"), sideEffect) is "hello151"
hi!
PASS foo(new String("hello"), sideEffect) is "hello152"
hi!
PASS foo(new String("hello"), sideEffect) is "hello153"
hi!
PASS foo(new String("hello"), sideEffect) is "hello154"
hi!
PASS foo(new String("hello"), sideEffect) is "hello155"
hi!
PASS foo(new String("hello"), sideEffect) is "hello156"
hi!
PASS foo(new String("hello"), sideEffect) is "hello157"
hi!
PASS foo(new String("hello"), sideEffect) is "hello158"
hi!
PASS foo(new String("hello"), sideEffect) is "hello159"
hi!
PASS foo(new String("hello"), sideEffect) is "hello160"
hi!
PASS foo(new String("hello"), sideEffect) is "hello161"
hi!
PASS foo(new String("hello"), sideEffect) is "hello162"
hi!
PASS foo(new String("hello"), sideEffect) is "hello163"
hi!
PASS foo(new String("hello"), sideEffect) is "hello164"
hi!
PASS foo(new String("hello"), sideEffect) is "hello165"
hi!
PASS foo(new String("hello"), sideEffect) is "hello166"
hi!
PASS foo(new String("hello"), sideEffect) is "hello167"
hi!
PASS foo(new String("hello"), sideEffect) is "hello168"
hi!
PASS foo(new String("hello"), sideEffect) is "hello169"
hi!
PASS foo(new String("hello"), sideEffect) is "hello170"
hi!
PASS foo(new String("hello"), sideEffect) is "hello171"
hi!
PASS foo(new String("hello"), sideEffect) is "hello172"
hi!
PASS foo(new String("hello"), sideEffect) is "hello173"
hi!
PASS foo(new String("hello"), sideEffect) is "hello174"
hi!
PASS foo(new String("hello"), sideEffect) is "hello175"
hi!
PASS foo(new String("hello"), sideEffect) is "hello176"
hi!
PASS foo(new String("hello"), sideEffect) is "hello177"
hi!
PASS foo(new String("hello"), sideEffect) is "hello178"
hi!
PASS foo(new String("hello"), sideEffect) is "hello179"
hi!
PASS foo(new String("hello"), sideEffect) is "hello180"
hi!
PASS foo(new String("hello"), sideEffect) is "hello181"
hi!
PASS foo(new String("hello"), sideEffect) is "hello182"
hi!
PASS foo(new String("hello"), sideEffect) is "hello183"
hi!
PASS foo(new String("hello"), sideEffect) is "hello184"
hi!
PASS foo(new String("hello"), sideEffect) is "hello185"
hi!
PASS foo(new String("hello"), sideEffect) is "hello186"
hi!
PASS foo(new String("hello"), sideEffect) is "hello187"
hi!
PASS foo(new String("hello"), sideEffect) is "hello188"
hi!
PASS foo(new String("hello"), sideEffect) is "hello189"
hi!
PASS foo(new String("hello"), sideEffect) is "hello190"
hi!
PASS foo(new String("hello"), sideEffect) is "hello191"
hi!
PASS foo(new String("hello"), sideEffect) is "hello192"
hi!
PASS foo(new String("hello"), sideEffect) is "hello193"
hi!
PASS foo(new String("hello"), sideEffect) is "hello194"
hi!
PASS foo(new String("hello"), sideEffect) is "hello195"
hi!
PASS foo(new String("hello"), sideEffect) is "hello196"
hi!
PASS foo(new String("hello"), sideEffect) is "hello197"
hi!
PASS foo(new String("hello"), sideEffect) is "hello198"
hi!
PASS foo(new String("hello"), sideEffect) is "hello199"
PASS successfullyParsed is true

TEST COMPLETE
